Add numerous impls of N-point crossover
This provides four distinct implementations of N-point crossover for benchmarking purposes. Three of these are about the same, and significantly faster than the fourth (`multi_swap`). `multi_swap` does appear to have the O(N^2) behavior that we suspected it might, and is thus noticeably slower for larger numbers of crossover points.