Change all the N-point XOs to use Floyd's selection algorithm
This changes all our N-point XO implementations to use our custom sampling using Floyd's algorithm instead of using `choose_multiple_fill()`, which is demonstrably very slow (at the moment).

This removed `NPointXoPrimitiveFloyds` because that is now the same as `NPointXoPrimitive`. We also had to remove the benchmark call on `NPointXoPrimitiveFloyds`.